Fixed build on GOARCH 386

+++ stat_times_386.go
@@ -1,4 +1,4 @@
-// +build !windows,!darwin,!freebsd,!386
+// +build !windows,!darwin,!freebsd,386
 
 // TODO: add more runtimes
 
@@ -14,8 +14,8 @@ func getTimeSpec(info os.FileInfo) timespec {
 	stat := info.Sys().(*syscall.Stat_t)
 	times := timespec{
 		Mtime: info.ModTime(),
-		Atime: time.Unix(stat.Atim.Sec, stat.Atim.Nsec),
-		Ctime: time.Unix(stat.Ctim.Sec, stat.Ctim.Nsec),
+		Atime: time.Unix(int64(stat.Atim.Sec), int64(stat.Atim.Nsec)),
+		Ctime: time.Unix(int64(stat.Ctim.Sec), int64(stat.Ctim.Nsec)),
 	}
 	return times
 }
